projects_toshare/wildbot/requirements.txt
discord.py
google-generativeai
python-dotenv